gui for retention basin and river section apps and their cooperation

+++ src/main/java/org/wilczewski/retentionbasin/IRetentionBasin.java
@@ -2,7 +2,7 @@ package org.wilczewski.retentionbasin;
 
 public interface IRetentionBasin {
     int getWaterDischarge();
-    long getFillingPercentage();
+    double getFillingPercentage();
     void setWaterDischarge(int waterDischarge);
     void setWaterInflow(int waterInflow, int port);
     void assignRiverSection(int port, String host);
